backend/requirements.txt
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Procesamiento de documentos
python-docx>=1.1.0
striprtf>=0.0.26
pywin32>=306; platform_system == "Windows"

# Cliente OpenAI
openai>=1.3.0
requests>=2.31.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.4.0
